# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ClusterPlan": {
        "additionalProperties": false,
        "description": "Ordered visit IDs per truck for one cluster; the shape of a solve artifact's routes.",
        "properties": {
          "cluster_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Cluster Id",
            "type": "string"
          },
          "routes": {
            "items": {
              "items": {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              "maxItems": 10000,
              "minItems": 1,
              "type": "array"
            },
            "maxItems": 10000,
            "minItems": 1,
            "title": "Routes",
            "type": "array"
          }
        },
        "required": [
          "cluster_id",
          "routes"
        ],
        "title": "ClusterPlan",
        "type": "object"
      },
...
      "Depot": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "lat": {
            "maximum": 90,
            "minimum": -90,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180,
            "minimum": -180,
            "title": "Lon",
            "type": "number"
          }
        },
        "required": [
          "id",
          "label",
          "lat",
          "lon"
        ],
        "title": "Depot",
        "type": "object"
      },
...
      "EvaluateRequest": {
        "additionalProperties": false,
        "description": "One cluster of a completed run, as its artifacts recorded it, plus the plans to evaluate.\n\n``cluster``, ``problem`` and ``travel`` are that cluster's entries of the clustering, problem\nand (decoded) travel artifacts; ``visits`` are the aggregation's visits of the cluster.\n``reference`` is normally the run's optimized routes, evaluated alongside for comparison.",
        "properties": {
          "cluster": {
            "additionalProperties": true,
            "title": "Cluster",
            "type": "object"
          },
          "plan": {
            "$ref": "#/components/schemas/ClusterPlan"
          },
          "problem": {
            "additionalProperties": true,
            "title": "Problem",
            "type": "object"
          },
          "reference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClusterPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "scenario": {
            "$ref": "#/components/schemas/ScenarioDocument"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "settings": {
            "$ref": "#/components/schemas/RunSettings"
          },
          "travel": {
            "additionalProperties": true,
            "title": "Travel",
            "type": "object"
          },
          "travel_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Snapshot"
          },
          "visits": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Visits",
            "type": "array"
          }
        },
        "required": [
          "scenario",
          "settings",
          "cluster",
          "problem",
          "travel",
          "visits",
          "plan"
        ],
        "title": "EvaluateRequest",
        "type": "object"
      },
      "EvaluateResponse": {
        "additionalProperties": false,
        "properties": {
          "cluster_id": {
            "title": "Cluster Id",
            "type": "string"
          },
          "evaluator_version": {
            "default": "fillrate-evaluate/1",
            "title": "Evaluator Version",
            "type": "string"
          },
          "manual": {
            "$ref": "#/components/schemas/PlanEvaluation"
          },
          "reference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanEvaluation"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "cluster_id",
          "manual"
        ],
        "title": "EvaluateResponse",
        "type": "object"
      },
...
      "GeometryLeg": {
        "additionalProperties": false,
        "properties": {
          "coordinates": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordinates"
          },
          "delta_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta M"
          },
          "delta_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta S"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "from_id": {
            "title": "From Id",
            "type": "string"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "matrix_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matrix M"
          },
          "matrix_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matrix S"
          },
          "notable": {
            "default": false,
            "title": "Notable",
            "type": "boolean"
          },
          "relative_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative M"
          },
          "relative_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative S"
          },
          "route_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route M"
          },
          "route_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route S"
          },
          "status": {
            "enum": [
              "ok",
              "same_location",
              "no_route",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "to_id": {
            "title": "To Id",
            "type": "string"
          }
        },
        "required": [
          "index",
          "from_id",
          "to_id",
          "status"
        ],
        "title": "GeometryLeg",
        "type": "object"
      },
      "GeometryStop": {
        "additionalProperties": false,
        "description": "A node of the truck's physical sequence: the depot, then each visit's location in order.",
        "properties": {
          "id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          }
        },
        "required": [
          "id",
          "lat",
          "lon"
        ],
        "title": "GeometryStop",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InventoryItem": {
        "additionalProperties": false,
        "properties": {
          "available_pieces": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "title": "Available Pieces",
            "type": "integer"
          },
          "product_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Product Id",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "available_pieces"
        ],
        "title": "InventoryItem",
        "type": "object"
      },
...
      "Order": {
        "additionalProperties": false,
        "properties": {
          "customer_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Customer Id"
          },
          "id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLine"
            },
            "minItems": 1,
            "title": "Lines",
            "type": "array"
          },
          "location_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Location Id",
            "type": "string"
          },
          "order_date": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Order Date",
            "type": "string"
          },
          "priority": {
            "default": 1,
            "maximum": 100,
            "minimum": 1,
            "title": "Priority",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "location_id",
          "order_date",
          "lines"
        ],
        "title": "Order",
        "type": "object"
      },
...
      "PlanEvaluation": {
        "additionalProperties": false,
        "properties": {
          "metrics": {
            "$ref": "#/components/schemas/PlanMetrics"
          },
          "trucks": {
            "items": {
              "$ref": "#/components/schemas/TruckSummary"
            },
            "title": "Trucks",
            "type": "array"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/PlanViolation"
            },
            "title": "Violations",
            "type": "array"
          }
        },
        "required": [
          "valid",
          "violations",
          "metrics",
          "trucks"
        ],
        "title": "PlanEvaluation",
        "type": "object"
      },
...
      "Product": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "linear_feet_per_piece": {
            "minimum": 1,
            "title": "Linear Feet Per Piece",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "label",
          "linear_feet_per_piece"
        ],
        "title": "Product",
        "type": "object"
      },
...
      "RouteGeometryRequest": {
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "snapshot": {
            "additionalProperties": true,
            "title": "Snapshot",
            "type": "object"
          },
          "snapshot_id": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Snapshot Id",
            "type": "string"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/GeometryStop"
            },
            "maxItems": 1000,
            "minItems": 2,
            "title": "Stops",
            "type": "array"
          },
          "truck_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Truck Id",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "snapshot",
          "truck_id",
          "stops"
        ],
        "title": "RouteGeometryRequest",
        "type": "object"
      },
      "RouteGeometryResponse": {
        "additionalProperties": false,
        "properties": {
          "chunks": {
            "additionalProperties": true,
            "title": "Chunks",
            "type": "object"
          },
          "geometry_version": {
            "default": "fillrate-route-geometry/1",
            "title": "Geometry Version",
            "type": "string"
          },
          "kind": {
            "const": "valhalla_road",
            "default": "valhalla_road",
            "title": "Kind",
            "type": "string"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/GeometryLeg"
            },
            "title": "Legs",
            "type": "array"
          },
          "note": {
            "default": "Valhalla's route for the same legs, shown for display. The plan was optimized on the recorded travel matrix; these paths do not prove which roads the solver used.",
            "title": "Note",
            "type": "string"
          },
          "provider": {
            "additionalProperties": true,
            "title": "Provider",
            "type": "object"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "truck_id": {
            "title": "Truck Id",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "truck_id",
          "provider",
          "legs",
          "chunks",
          "summary"
        ],
        "title": "RouteGeometryResponse",
        "type": "object"
      },
      "RunSettings": {
        "additionalProperties": false,
        "properties": {
          "allocation_objective": {
            "default": "revenue",
            "enum": [
              "revenue",
              "priority_then_revenue"
            ],
            "title": "Allocation Objective",
            "type": "string"
          },
          "allocation_strategy": {
            "default": "order_date_then_value",
            "enum": [
              "order_date_then_value",
              "first_come",
              "priority",
              "proportional",
              "optimized"
            ],
            "title": "Allocation Strategy",
            "type": "string"
          },
          "allocation_time_limit_s": {
            "default": 10,
            "exclusiveMinimum": 0,
            "maximum": 300,
            "title": "Allocation Time Limit S",
            "type": "number"
          },
          "auto_k_cap": {
            "default": 25,
            "maximum": 100,
            "minimum": 1,
            "title": "Auto K Cap",
            "type": "integer"
          },
          "cluster_circuity": {
            "default": 1.2,
            "maximum": 5,
            "minimum": 1,
            "title": "Cluster Circuity",
            "type": "number"
          },
          "cluster_strategy": {
            "default": "kmeans",
            "enum": [
              "kmeans",
              "h3",
              "none"
            ],
            "title": "Cluster Strategy",
            "type": "string"
          },
          "cost_per_mile_cents": {
            "anyOf": [
              {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Cost Per Mile Cents"
          },
          "cost_per_truck_cents": {
            "anyOf": [
              {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Cost Per Truck Cents"
          },
          "excluded_line_ids": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 25000,
            "title": "Excluded Line Ids",
            "type": "array"
          },
          "fulfillment_policy": {
            "default": "piece",
            "enum": [
              "piece",
              "whole_order"
            ],
            "title": "Fulfillment Policy",
            "type": "string"
          },
          "h3_resolution": {
            "default": 2,
            "maximum": 15,
            "minimum": 0,
            "title": "H3 Resolution",
            "type": "integer"
          },
          "inventory_percent": {
            "default": 100,
            "maximum": 1000,
            "minimum": 0,
            "title": "Inventory Percent",
            "type": "integer"
          },
          "k": {
            "anyOf": [
              {
                "maximum": 1000,
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "K"
          },
          "kmeans_n_init": {
            "default": 10,
            "maximum": 100,
            "minimum": 1,
            "title": "Kmeans N Init",
            "type": "integer"
          },
          "kmeans_seed": {
            "default": 0,
            "maximum": 2147483647,
            "minimum": 0,
            "title": "Kmeans Seed",
            "type": "integer"
          },
          "max_cluster_diameter_m": {
            "anyOf": [
              {
                "maximum": 20000000,
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Max Cluster Diameter M"
          },
          "max_leg_m": {
            "default": 804672,
            "maximum": 20000000,
            "minimum": 1,
            "title": "Max Leg M",
            "type": "integer"
          },
          "max_stops": {
            "default": 500,
            "maximum": 10000,
            "minimum": 1,
            "title": "Max Stops",
            "type": "integer"
          },
          "objective": {
            "default": "trucks_then_distance",
            "enum": [
              "trucks_then_distance",
              "weighted_distance",
              "cost"
            ],
            "title": "Objective",
            "type": "string"
          },
          "preflight": {
            "$ref": "#/components/schemas/PreflightPolicy"
          },
          "respect_order_date": {
            "default": false,
            "title": "Respect Order Date",
            "type": "boolean"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "solver_max_iterations": {
            "anyOf": [
              {
                "maximum": 10000000,
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Solver Max Iterations"
          },
          "solver_seed": {
            "default": 0,
            "maximum": 2147483647,
            "minimum": 0,
            "title": "Solver Seed",
            "type": "integer"
          },
          "solver_time_limit_s": {
            "default": 10,
            "exclusiveMinimum": 0,
            "maximum": 300,
            "title": "Solver Time Limit S",
            "type": "number"
          },
          "trailer_capacity": {
            "default": 5300,
            "maximum": 1000000,
            "minimum": 1,
            "title": "Trailer Capacity",
            "type": "integer"
          },
          "travel_circuity": {
            "default": 1.2,
            "maximum": 5,
            "minimum": 1,
            "title": "Travel Circuity",
            "type": "number"
          },
          "travel_snapshot_id": {
            "anyOf": [
              {
                "pattern": "^[a-f0-9]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Travel Snapshot Id"
          },
          "warm_start": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WarmStartSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "weighted_truck_penalty_m": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Weighted Truck Penalty M"
          }
        },
        "title": "RunSettings",
        "type": "object"
      },
...
      "ScenarioDocument": {
        "additionalProperties": false,
        "properties": {
          "depot": {
            "$ref": "#/components/schemas/Depot"
          },
          "inventory": {
            "items": {
              "$ref": "#/components/schemas/InventoryItem"
            },
            "title": "Inventory",
            "type": "array"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/Location"
            },
            "title": "Locations",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "title": "Orders",
            "type": "array"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "title": "Products",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "time_model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "depot",
          "products",
          "locations",
          "orders",
          "inventory"
        ],
        "title": "ScenarioDocument",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Fillrate optimizer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/evaluate": {
      "post": {
        "description": "Manual plan evaluation (spec \u00a710): bounded, synchronous CPU work in FastAPI's thread\npool, never on the event loop. Only Next.js calls it, with the worker bearer token.",
        "operationId": "post_evaluate_evaluate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Content"
          }
        },
        "summary": "Post Evaluate"
      }
    },
...
    "/route-geometry": {
      "post": {
        "description": "Road geometry for one inspected truck (spec \u00a74, \u00a77): bounded synchronous Valhalla `/route`\ncalls in the thread pool. Only Next.js calls it, with the worker bearer token.",
        "operationId": "post_route_geometry_route_geometry_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteGeometryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteGeometryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Content"
          },
          "502": {
            "description": "Bad Gateway"
          }
        },
        "summary": "Post Route Geometry"
      }
    },
    "/route-geometry/context": {
      "get": {
        "description": "This deployment's Valhalla identity (no endpoint), for the web's eligibility check.",
        "operationId": "get_route_geometry_context_route_geometry_context_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Route Geometry Context Route Geometry Context Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Route Geometry Context"
      }
    }
  }
}